api/requirements.txt
Django
djangorestframework
django-filter
requests==2.32.3
